services/ideas-service/pyproject.toml
[project]
name = "ideas-service"
version = "0.1.0"
description = "Keiko Ideas Service"
authors = [
    {name = "Keiko Team", email = "team@keiko.ai"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "azure-cosmos>=4.8.0",
    "azure-search-documents>=11.4.0",
    "azure-identity>=1.18.0",
    "openai>=1.47.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
